#40 whole lot of changes, db stuff, cache improvements, streamlining slash command updating to be done on application load for easier testing

+++ db/models/channel.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, DateTime, ForeignKey
+from sqlalchemy import create_engine, Column, Integer, String, DateTime, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 
 from . import Base
@@ -7,7 +7,7 @@ from . import Base
 class Channel(Base):
 
     __tablename__ = 'channels'
-    id = Column(Integer, primary_key=True)
-    guild_id = Column(Integer, ForeignKey('guilds.id'))
-    mirror_to_channel_id = Column(Integer, ForeignKey('channels.id'), nullable=True)
+    id = Column(BigInteger, primary_key=True)
+    guild_id = Column(BigInteger, ForeignKey('guilds.id'))
+    mirror_to_channel_id = Column(BigInteger, ForeignKey('channels.id'), nullable=True)
     guild = relationship("Guild", back_populates="channels")
