+++ db/models/guild.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, DateTime
+from sqlalchemy import create_engine, Column, Integer, String, DateTime, BigInteger
 from sqlalchemy.orm import relationship
 
 from . import Base
@@ -7,5 +7,5 @@ from . import Base
 class Guild(Base):
 
     __tablename__ = 'guilds'
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     channels = relationship('Channel', back_populates='guild')
